Rescale norm constraint
Mean square weight is more interpretable than sum of squared weights,
because it doesn't depend on the sample size. We can also interpret
the mean square weight as the variance inflation factor, since the
mean square weight equals the ratio of the achieved variance to the
variance of the unweighted average.