Constrain mean weight
Make it easy for the user to constrain the mean weight, e.g. to 1.
This creates 3 possibilities: a true weighted average of the
responses (how people expect weights to behave, and the new default);
some other constraint (such as maintaining consistency with the
baseline weights); no constraint. No constraint is the legacy
behavior, and is worth considering since an extra constraint cannot
decrease the distance from the baseline.