Fix bug with EfficientFrontier
When we instantiate Rake without a constraint on the norm of the
weights, the phase I solver will be an EqualityWithBoundsSolver, but
then when we want to trace out the efficient frontier, we need an
EqualityWithBoundsAndNormConstraintSolver. I had been assuming the
Rake phase 1 solver was the latter, but in this case it's the former,
so I wrap the solver to get the desired behavior.